re-enable the test I disabled before and move the test from integration-tests to client
