+++ chain/client/src/tests/mod.rs
@@ -3,4 +3,5 @@ mod catching_up;
 mod chunks_management;
 mod consensus;
 mod cross_shard_tx;
+mod process_blocks;
 mod query_client;
